% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addSubjectICV}
\alias{addSubjectICV}
\alias{addSubjectTBV}
\title{Utility functions}
\usage{
addSubjectICV(subjectData)

addSubjectTBV(subjectData)
}
\arguments{
\item{subjectData}{output from \code{readSubject} in the form of a dataframe}
}
\value{
data frame with ICV or TBV added
}
\description{
takes in a subject data frame and adds an ICV variable
}
\details{
addSubjectICV
}
