% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addSubjectTBV}
\alias{addSubjectTBV}
\title{Adds an ICV measure to subejct data}
\usage{
addSubjectTBV(subjectData)
}
\arguments{
\item{subjectData}{output from \code{readSubject} in the form of a dataframe}
}
\description{
Adds an ICV measure to subejct data
}
