% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixBasalForebrain}
\alias{fixBasalForebrain}
\title{A function for fixing the basal forebrain labeling error}
\usage{
fixBasalForebrain(subjectData)
}
\arguments{
\item{subjectData}{output from \code{readSubject} in the form of a dataframe}
}
\description{
A function for fixing the basal forebrain labeling error
}
