% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixTarget2}
\alias{fixTarget2}
\title{Fixes a rawid label error}
\usage{
fixTarget2(subjectData)
}
\arguments{
\item{subjectData}{output from \code{readSubject} in the form of a dataframe}
}
\description{
Fixes a rawid label error
}
