% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{A function for compositional principal component analysis of ROIs 
across age}
\usage{
PCA(data, infodir, Type = 1, Level = 1)
}
\arguments{
\item{data}{use \code{\link{readFileList}} 
function to get MRI data frame}

\item{infodir}{path to lookup table of patient info}

\item{Type}{optinal,type of MRI sequences(T1-weighted, T2-weighted, etc.), 
default = 1}

\item{Level}{optional,level of MRI, default = 1}
}
\description{
A function for compositional principal component analysis of ROIs 
across age
}
\author{
Luchao Qi
}
