% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAvisualization.R
\name{PCAvisualization}
\alias{PCAvisualization}
\title{A function for visualization of principal component rotations}
\usage{
PCAvisualization(pca.result, PC = 1)
}
\arguments{
\item{pca.result}{should use function MRIPCA to get a list of pca result}

\item{PC}{index of PC, default = 1}
}
\description{
A function for visualization of principal component rotations
}
\author{
Luchao Qi
}
