% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileList.R
\name{readFileList}
\alias{readFileList}
\title{A function for reading a list of T1 volumetric mricloud data}
\usage{
readFileList(fileList, levelList = 5, typeList = 1)
}
\arguments{
\item{fileList}{a textlist of files}

\item{levelList}{an optional list of levels to include, needs equal length as \code{typeList}}

\item{typeList}{an optional list of types to include, needs equal length to \code{levelList}}
}
\description{
A function for reading in a collection of files of T1 volumetric output
from MRIcloud. The files must be a list of files
}
