% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSubject.R
\name{readSubject}
\alias{readSubject}
\title{readSubject}
\usage{
readSubject(fileloc)
}
\arguments{
\item{fileloc}{the location of the volumetric processed output text file}
}
\description{
A utility function for parsing the output of MRIcloud volumetric output
}
\details{
This function takes in a filename that consist of volumetric output from
MRIcloud. It then parses it into a list of dataframes of the various combinations 
of levels and outputs. It requires the full path of to the file as text input.
}
\author{
Brian Caffo
}
\keyword{MRI,}
\keyword{structural}
