% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSubjectDf}
\alias{readSubjectDf}
\title{reads in a subject as a df}
\usage{
readSubjectDf(filename)
}
\arguments{
\item{filename}{name of the file to read in}
}
\description{
reads in a subject as a df
}
