% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadROIs.R
\name{spreadROIs}
\alias{spreadROIs}
\title{Converts the long data format to wide for ROI data}
\usage{
spreadROIs(dat)
}
\arguments{
\item{dat}{the output from \code{readFileList}}
}
\description{
this function just used tidyverse spread for output specifically from \code{readFileList}
only does this for volume and only if you haven't modified the data frame. should be ok if you select type and level
}
