% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject2df.R
\name{subject2df}
\alias{subject2df}
\title{Takes the output from \code{readSubject} and converts it to a dataframe}
\usage{
subject2df(subjectData)
}
\description{
Takes a subject from readSubject and converts it into a single
dataframe
}
