% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileList.R
\name{readFileList}
\alias{readFileList}
\title{A function for reading a list of T1 volumetric mricloud data}
\usage{
readFileList(fileList, idList = NULL)
}
\arguments{
\item{fileList}{a textlist of files}

\item{idList}{an optional list of ids, otherwise is 1 : length(fileList)}
}
\description{
A function for reading in a collection of files of T1 volumetric output
from MRIcloud. The files must be a list of files
}
